==> Analysis of AI and Student-written Code/1. Clone Student Repos.ipynb <==
import gitlab
from git import Repo
import os
import pandas as pd
import time
import shutil
import stat

## Downloading the student solutions

So that the student solutions remain anonymised, this notebook cannot be re-run. The anonymised student solutions are stored in: ss21_raw_solutions/student_solutions.

Import the list of repo URLS for each milestone:

milestone_lists = [pd.read_csv("student_repos_list/st2m0.csv"),
                   pd.read_csv("student_repos_list/st2m1.csv"),
                   pd.read_csv("student_repos_list/st2m2.csv"),
                   pd.read_csv("student_repos_list/st2m3.csv"),
                   pd.read_csv("student_repos_list/st2m4.csv")]
milestone_names = ["m0", "m1", "m2", "m3", "m4"]

Clone every solution repo:

token = os.environ.get("GITLAB_ACCESS_TOKEN")
gl = gitlab.Gitlab('https://git.st.archi-lab.io/', private_token=token)

for milestone, milestone_name in zip(milestone_lists, milestone_names):
    for repo, student_name in zip(milestone["Code Repo"], milestone["Student"]):
        repo_url = repo + ".git"
        Repo.clone_from(repo_url, f"ss21_raw_solutions/student_solutions/{milestone_name}_{student_name.strip()}")
        time.sleep(0.06)

Check the repos were cloned correctly:

# generate all of the unique repos that were cloned
# list the milestone and student names from the directory
cloned_repos = pd.Series(os.listdir("ss21_raw_solutions/student_solutions")).str.split("_", expand=True)
cloned_repos = cloned_repos.rename(columns={0:"Milestone", 1:"Student"}).sort_values(["Milestone", "Student"]).reset_index(drop=True)

# generate all of the unique repos from the list of URLS
# add the milestone name to student name \ for every milestone
URL_repos = pd.concat([pd.concat([pd.Series([milestone_name]*len(milestone["Student"])), milestone["Student"].str.strip()], axis=1)\
                       for milestone_name, milestone in zip(milestone_names, milestone_lists)])
URL_repos = URL_repos.rename(columns={0:"Milestone"}).sort_values(["Milestone", "Student"]).reset_index(drop=True)

cloned_repos.equals(URL_repos)

## Downloading the students' hidden tests

Cloning the test repos:

m0_students = milestone_lists[0]["Student"]
m0_tests =  milestone_lists[0]["Test Repo"]

for student_name, test_repo in zip(m0_students, m0_tests):
    repo_url = test_repo + ".git"
    Repo.clone_from(repo_url, f"ss21_raw_solutions/student_tests/test_{student_name.strip()}")
    time.sleep(0.06)

Moving the hidden test files into the student solutions:

with open("student_repos_list/student_anonymised_dict.json", 'r') as json_file:
    id_to_student = json.load(json_file)

for student_id, student_name in id_to_student.items():
    test_path = f"ss21_raw_solutions/student_tests/test_{student_name}/src/test/java/thkoeln/st/st2praktikum/exercise/Exercise0HiddenTests.java"
    solution_path = f"ss21_raw_solutions/student_solutions/m0_{student_id}/src/test/java/thkoeln/st/st2praktikum/exercise"
    shutil.copy(test_path, solution_path)
