==> Analysis of AI and Student-written Code/3b. Generate Embeddings.ipynb <==
import pandas as pd
import numpy as np
import scipy as sp
import tiktoken
import openai
import time
from openai.embeddings_utils import get_embedding

Getting the embeddings from the api:

def get_embeddings(code):
    embeddings = []
    for solution in code:
        embeddings.append(get_embedding(solution, engine=embedding_model))
        time.sleep(1)
    return embeddings

### Generating embeddings for the ST2 M0 solutions:

m0_data = pd.read_csv("ss21_processed_solutions/m0_data.csv")
m0_code = list(m0_data["code"])

Calculating how many tokens will be embedded:

embedding_model = "text-embedding-ada-002"
embedding_encoding = "cl100k_base"

encoding = tiktoken.get_encoding(embedding_encoding)
num_tokens = [len(encoding.encode(code)) for code in m0_code]

print("largest number of tokens:", max(num_tokens))
print("total number of tokens:", sum(num_tokens))
print("total cost: $", sum(num_tokens)/1000*0.0004)

m0_embeddings = get_embeddings(m0_code)

m0_data["embedding"] = m0_embeddings
m0_data = pd.concat([m0_data, pd.DataFrame(m0_embeddings)], axis=1)

m0_data.to_csv("ss21_processed_solutions/m0_embedding_data.csv", index=False)

### Generating embeddings for the modified solutions:

optimised_data = pd.read_csv("ss21_processed_solutions/optimised_data.csv")
optimised_code = list(optimised_data["code"])

optimised_embeddings = get_embeddings(optimised_code)

optimised_data["embedding"] = optimised_embeddings
optimised_data = pd.concat([optimised_data.reset_index(drop=True), pd.DataFrame(optimised_embeddings)], axis=1)

optimised_data.to_csv("ss21_processed_solutions/optimised_embedding_data.csv", index=False)
