==> Analysis of AI and Student-written Code/4b. Modelling (Embedding Approach).ipynb <==
import pandas as pd
import numpy as np
import scipy as sp
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.metrics import accuracy_score, precision_recall_curve

# Modelling with embedding approach

embed_data = pd.read_csv("ss21_processed_solutions/m0_embedding_data.csv")
embed_data["binary_source"] = embed_data["source"].apply(lambda row: row if row=="student" else "ai")

embed_features = list(embed_data.columns[6:-1])

X_train_embed, X_test_embed, y_train_embed, y_test_embed = train_test_split(embed_data[embed_features].values, embed_data["binary_source"], test_size=0.3, random_state=0)

embed_model = GridSearchCV(RandomForestClassifier(random_state=0), param_grid={"n_estimators":np.arange(100,200,10),
                                                                             "max_depth":[10, 20, 30, None], 
                                                                             "max_features":[1, 0.9, 0.8, "sqrt", "log2"]}) 
embed_model.fit(X_train_embed, y_train_embed)
embed_model_results = pd.DataFrame(embed_model.cv_results_)
embed_model_results[embed_model_results["rank_test_score"]==1].head()

embed_model = RandomForestClassifier(random_state=0, max_depth=10, max_features=0.9, n_estimators=190)
embed_model.fit(X_train_embed, y_train_embed)
print("Accuracy on test set:", np.round(accuracy_score(y_test_embed, embed_model.predict(X_test_embed)), 3)*100, "%")

misclassified = y_test_embed != embed_model.predict(X_test_embed)
(embed_data.loc[misclassified.index[misclassified]]["source"].value_counts() / embed_data.loc[y_test_embed.index]["source"].value_counts()).dropna() * 100

## Classifying each source of data

for source in embed_data["source"].unique()[:-1]:
    embed_model = RandomForestClassifier(random_state=0, max_depth=10, max_features=0.9, n_estimators=190)
    embed_model.fit(embed_data[embed_data["source"] != source][embed_features], embed_data[embed_data["source"] != source]["binary_source"])
    print(f"Accuracy on classifying {source}:", np.round(accuracy_score(embed_data[embed_data["source"] == source]["binary_source"], 
                                                            embed_model.predict(embed_data[embed_data["source"] == source][embed_features])), 3)*100, "%")

## Testing the optimised embedded solutions on the model

optimised_data = pd.read_csv("ss21_processed_solutions/optimised_embedding_data.csv")
optimised_data["binary_source"] = optimised_data["source"].apply(lambda row: row if row=="student" else "ai")

full_model = RandomForestClassifier(random_state=0, max_depth=10, max_features=0.9, n_estimators=190)
full_model.fit(embed_data[embed_features].values, embed_data["binary_source"])

predictions = pd.concat([pd.DataFrame(full_model.predict_proba(optimised_data[embed_features].values), columns=["Probability AI", "Probability Human"]),
                        optimised_data.reset_index(drop=True)], axis=1)
predictions.iloc[:,:8].sort_values("Probability Human")
